+++ tests/test_base_components.py
@@ -8,7 +8,6 @@ from sympl import (
     ComponentMissingOutputError, ComponentExtraOutputError,
     InvalidStateError
 )
-import warnings
 
 def same_list(list1, list2):
     return (len(list1) == len(list2) and all(
@@ -449,24 +448,18 @@ class DiagnosticTestBase():
         with self.assertRaises(InvalidPropertyDictError):
             self.get_component(diagnostic_properties=diagnostic_properties)
 
-    def test_diagnostic_warns_when_units_incompatible_with_input(self):
+    def test_diagnostic_raises_when_units_incompatible_with_input(self):
         input_properties = {
             'diag1': {'units': 'km', 'dims': ['dim1', 'dim2']}
         }
         diagnostic_properties = {
             'diag1': {'units': 'seconds', 'dims': ['dim1', 'dim2']}
         }
-        with warnings.catch_warnings(record=True) as w:
+        with self.assertRaises(InvalidPropertyDictError):
             self.get_component(
                 input_properties=input_properties,
                 diagnostic_properties=diagnostic_properties
             )
-            assert len(w) == 1
-            assert issubclass(w[-1].category, UserWarning)
-            assert 'units' in str(w[-1].message)
-            assert 'diag1' in str(w[-1].message)
-            assert 'seconds' in str(w[-1].message)
-            assert 'km' in str(w[-1].message)
 
     def test_diagnostic_requires_correct_number_of_dims(self):
         input_properties = {
@@ -555,6 +548,7 @@ class DiagnosticTestBase():
         assert len(diagnostics['output1'].dims) == 1
         assert 'dim1' in diagnostics['output1'].dims
         assert 'units' in diagnostics['output1'].attrs
+        assert len(diagnostics['output1'].attrs) == 1
         assert diagnostics['output1'].attrs['units'] == 'm'
         assert np.all(diagnostics['output1'].values == np.ones([10]))
 
@@ -771,24 +765,18 @@ class PrognosticTests(unittest.TestCase, InputTestBase):
         instance = MyPrognostic()
         assert isinstance(instance, Prognostic)
 
-    def test_tendency_warns_when_units_incompatible_with_input(self):
+    def test_tendency_raises_when_units_incompatible_with_input(self):
         input_properties = {
             'input1': {'units': 'km', 'dims': ['dim1', 'dim2']}
         }
         tendency_properties = {
             'input1': {'units': 'degK/s', 'dims': ['dim1', 'dim2']}
         }
-        with warnings.catch_warnings(record=True) as w:
+        with self.assertRaises(InvalidPropertyDictError):
             self.get_component(
                 input_properties=input_properties,
                 tendency_properties=tendency_properties
             )
-            assert len(w) == 1
-            assert issubclass(w[-1].category, UserWarning)
-            assert 'units' in str(w[-1].message)
-            assert 'input1' in str(w[-1].message)
-            assert 'degK/s' in str(w[-1].message)
-            assert 'km' in str(w[-1].message)
 
     def test_two_components_are_not_instances_of_each_other(self):
         class MyPrognostic1(Prognostic):
@@ -865,7 +853,7 @@ class PrognosticTests(unittest.TestCase, InputTestBase):
     def test_tendency_uses_base_dims(self):
         input_properties = {'diag1': {'dims': ['dim1'], 'units': 'm'}}
         diagnostic_properties = {}
-        tendency_properties = {'diag1': {'units': 'm'}}
+        tendency_properties = {'diag1': {'units': 'm/s'}}
         diagnostic_output = {}
         tendency_output = {}
         self.component_class(
@@ -1045,6 +1033,7 @@ class PrognosticTests(unittest.TestCase, InputTestBase):
         assert len(tendencies['output1'].dims) == 1
         assert 'dim1' in tendencies['output1'].dims
         assert 'units' in tendencies['output1'].attrs
+        assert len(tendencies['output1'].attrs) == 1
         assert tendencies['output1'].attrs['units'] == 'm/s'
         assert np.all(tendencies['output1'].values == np.ones([10]))
 
@@ -1542,24 +1531,18 @@ class ImplicitTests(unittest.TestCase, InputTestBase, DiagnosticTestBase):
         instance = MyImplicit()
         assert isinstance(instance, Implicit)
 
-    def test_output_warns_when_units_incompatible_with_input(self):
+    def test_output_raises_when_units_incompatible_with_input(self):
         input_properties = {
             'input1': {'units': 'km', 'dims': ['dim1', 'dim2']}
         }
         output_properties = {
             'input1': {'units': 'degK', 'dims': ['dim1', 'dim2']}
         }
-        with warnings.catch_warnings(record=True) as w:
+        with self.assertRaises(InvalidPropertyDictError):
             self.get_component(
                 input_properties=input_properties,
                 output_properties=output_properties,
             )
-            assert len(w) == 1
-            assert issubclass(w[-1].category, UserWarning)
-            assert 'units' in str(w[-1].message)
-            assert 'input1' in str(w[-1].message)
-            assert 'degK' in str(w[-1].message)
-            assert 'km' in str(w[-1].message)
 
     def test_two_components_are_not_instances_of_each_other(self):
         class MyImplicit1(Implicit):
@@ -1804,7 +1787,7 @@ class ImplicitTests(unittest.TestCase, InputTestBase, DiagnosticTestBase):
         output_properties = {
             'output1': {
                 'dims': ['dim1'],
-                'units': 'm/s',
+                'units': 'm',
             }
         }
         diagnostic_output = {}
@@ -1830,7 +1813,7 @@ class ImplicitTests(unittest.TestCase, InputTestBase, DiagnosticTestBase):
         assert len(output['output1'].dims) == 1
         assert 'dim1' in output['output1'].dims
         assert 'units' in output['output1'].attrs
-        assert output['output1'].attrs['units'] == 'm/s'
+        assert output['output1'].attrs['units'] == 'm'
         assert np.all(output['output1'].values == np.ones([10]))
 
     def test_output_with_dims_from_input(self):
@@ -1843,7 +1826,7 @@ class ImplicitTests(unittest.TestCase, InputTestBase, DiagnosticTestBase):
         diagnostic_properties = {}
         output_properties = {
             'output1': {
-                'units': 'm/s',
+                'units': 'm',
             }
         }
         diagnostic_output = {}
@@ -1869,7 +1852,7 @@ class ImplicitTests(unittest.TestCase, InputTestBase, DiagnosticTestBase):
         assert len(output['output1'].dims) == 1
         assert 'dim1' in output['output1'].dims
         assert 'units' in output['output1'].attrs
-        assert output['output1'].attrs['units'] == 'm/s'
+        assert output['output1'].attrs['units'] == 'm'
         assert np.all(output['output1'].values == np.ones([10]))
 
     def test_tendencies_in_diagnostics_no_tendency(self):